change BD

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -5,12 +5,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.webjars.NotFoundException;
-import ru.skypro.homework.dto.Ad;
-import ru.skypro.homework.dto.Ads;
-import ru.skypro.homework.dto.CreateOrUpdateAd;
-import ru.skypro.homework.dto.ExtendedAd;
+import ru.skypro.homework.dto.*;
 import ru.skypro.homework.entity.AdEntity;
 import ru.skypro.homework.entity.ImageEntity;
+import ru.skypro.homework.entity.UserEntity;
 import ru.skypro.homework.mappers.AdMapper;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.ImageEntityRepository;
@@ -31,15 +29,17 @@ public class AdServiceImpl implements AdService {
     private final AdRepository repository;
     //private final ImageServiceImpl imageService;
     private final ImageEntityRepository imageEntityRepository;
+    private final UserServiceImpl userService;
 
     public Ad addAd(MultipartFile image, CreateOrUpdateAd properties) throws IOException {
         AdEntity adEntity = AdMapper.INSTANCE.createOrUpdateToAdEntity(properties);
         repository.save(adEntity);
-        uploadImageForAd(repository.save(adEntity).getId(),image);
+        uploadImageForAd(repository.save(adEntity).getId(), image);
 
         return AdMapper.INSTANCE.adEntityToAd(adEntity);
 
     }
+
     public void uploadImageForAd(Long id, MultipartFile image) throws IOException {
 
         AdEntity ad = findById(id);
@@ -62,10 +62,10 @@ public class AdServiceImpl implements AdService {
         imageEntityRepository.save(imageEntity);
         ad.setImageEntity(imageEntity);
     }
+
     private String getExtensions(String fileName) {
         return fileName.substring(fileName.lastIndexOf(".") + 1);
     }
-
 
 
     public Ads getAll() {
@@ -79,15 +79,20 @@ public class AdServiceImpl implements AdService {
         return repository.findById(id).orElseThrow();
     }
 
-    public void deleteAd(Integer id) {
-        repository.deleteById(id.longValue());
+    public boolean deleteAd(Integer id, String login) {
+        AdEntity adEntity = repository.findById(id.longValue()).get();
+        UserEntity userEntity = userService.findByLogin(login);
+        if ((adEntity.getUser().getLogin()).equals(login) || userEntity.getRole() == Role.ADMIN) {
+            repository.deleteById(id.longValue());
+            return true;}
+        else return false;
+        }
+
+        public ExtendedAd getById (Integer id){
+            Optional<AdEntity> ad = repository.findById(id.longValue());
+            if (ad.isPresent()) {
+                return AdMapper.INSTANCE.adEntityToExtendedAd(ad.get());
+            } else throw new NotFoundException("Отсутствует в базе");
+        }
 
     }
-
-    public ExtendedAd getById(Integer id) {
-        Optional<AdEntity> ad= repository.findById(id.longValue());
-        if (ad.isPresent()){return AdMapper.INSTANCE.adEntityToExtendedAd(ad.get());}
-        else throw new NotFoundException("Отсутствует в базе");
-    }
-
-}
